Cart class now stores list of product structs, instead of strings
These product structs are converted to just their ID when passing through JSON. This is done so that we are not a target of attack for modifying values on the output. For example, prices cannot be modified on out service in order to attack the rest of the ecommerce system.

Next I will need to replace the new placeholder in the html with new html for a product